Simplify code (pass 5)
